+++ .eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+	root: true,
+	extends: [
+		'./config/common.js',
+		'./config/ecmascript-7.js',
+		'./config/node.js',
+	]
+};